classify/chinese_classify/sogouNews_FlyAI/requirements.txt
pandas
beautifulsoup4
tensorflow-gpu
keras
flyai-gpu
numpy
jieba
